backend/pyproject.toml
[project]
name = "kayman"
version = "0.10.0"
description = "Kayman is the one-stop solution for personal finance"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
authors = [{ name = "Tomy Hsieh", email = "pypi@tomy.me" }]
dependencies = [
    "fastapi (>=0.115,<0.116)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic (>=2.0,<3.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "uvicorn[standard] (>=0.34,<0.35)",
    "psycopg2 (>=2.9.3,<3.0.0)",
    "alembic (>=1.8.1,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<3.5.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "loguru (>=0.6.0,<0.7.0)",
    "tzdata (>=2022.4,<2023.0)",
    "simplejson (>=3.17.6,<4.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pydantic-extra-types (>=2.10.2,<3.0.0)",
]

[project.urls]
repository = "https://github.com/tomy0000000/kayman"
documentation = "https://github.com/tomy0000000/kayman/wiki"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ipython = "^8.5.0"
mypy = ">=1.14.1,<1.15.0"
pytest = "^8.3.4"
httpx = "^0.28.1"
factory-boy = ">=3.3.3,<3.4.0"
pre-commit = "^4.1.0"
ruff = "^0.9.2"
types-simplejson = "^3.19.0.20241221"
types-python-jose = ">=3.4.0,<3.5.0"
pytest-cov = "^6.0.0"

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files.html
profile = "black"
src_paths = ["kayman"]

[tool.pytest.ini_options]
# https://docs.pytest.org/en/stable/reference/customize.html#configuration-file-formats
addopts = "--capture=no --cov=kayman --cov-report=term-missing --cov-report=xml --cov-report=html"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html
strict = true
exclude = ["venv", ".venv", "alembic", "kayman/tests"]

[[tool.mypy.overrides]]
# TODO: Remove this when factory boy supports typing
module = ["factory.*"]
ignore_missing_imports = true

[tool.ruff]
# https://docs.astral.sh/ruff/settings/
target-version = "py312" # Python 3.12
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "B008", # do not perform function calls in argument defaults
]
